requirements.txt
Flask==3.0.3
python-telegram-bot==20.7
httpx==0.25.2
anyio==4.3.0
